backend/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
black==25.1.0
cachetools==5.5.2
celery==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.44.0
colorama==0.4.6
dj-database-url==2.3.0
Django==5.1.6
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-debug-toolbar==5.2.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
dotenv==0.9.9
flake8==7.2.0
google-auth==2.40.1
gunicorn==23.0.0
h11==0.16.0
idna==3.10
isort==6.0.1
kombu==5.5.3
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.13.0
pyflakes==3.3.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==6.0.0
requests==2.32.3
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
